Refactor isFirstVisitAtom to store directory and remove unused hook
